Make clear need to maintain gauge object reference
Previously the example code could be misinterpreted as being usable as-is in arbitrary places in code.
This commit adds comments to the example code to clarify the need to maintain a reference to the object backing the Gauge.
It also attempts to make clear the updating of the backing object can happen in arbitrary places, using the reference to the backing object.

Resolves #107