Stackdriver resource type/labels and unique tagging docs (#191)
Stackdriver fails to write metrics if two application instances are trying to write identical metrics in the same interval. To make the meters unique, a label or labels unique to the instance should be configured either as common tags or as resource labels.

information about resource type and resource labels was missing from the docs previously. Links to the official docs.