Serve only docs from /docs for embedding in new website
We want these docs to be available at the same web address as they have been, so we do not break links on the internet to our docs. However, we are planning to launch a new website with a different build from a different source repository. Rather than bring the legacy of this source and build, we want to try embedding the built static website from here into the new website.

We should clearly indicate to visitors of the old docs that these are archived and will not be updated. We can link to the new docs so when users arrive from old links they know how to get to the new docs.